baselines/requirements.txt
# Baseline model requirements
monodepth2 @ git+https://github.com/nianticlabs/monodepth2.git
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
pillow>=7.0.0
matplotlib>=3.3.0
opencv-python>=4.4.0
psutil>=5.7.0